Laravel 5 update
Autoload fixes

Fixing the facade

small file rename

Readme update

Readme typo
