+++ composer.json
@@ -12,12 +12,12 @@
         }
     ],
     "require": {
-        "php": ">=5.3.0",
-        "illuminate/support": "~5.0",
+        "php": ">=5.4.0",
+        "illuminate/support": "5.0.x|5.1.x|5.2.x",
         "opentok/opentok": "2.2.x"
     },
     "autoload": {
-        "psr-0": {
+        "psr-4": {
             "Tomcorbett\\OpentokLaravel\\": "src/"
         }
     },
