+++ config/opentok.php
@@ -1,0 +1,6 @@
+<?php 
+
+return	[
+    'api_key' => 'API_KEY',
+    'api_secret' => 'API_SECRET',
+];
