--- src/Tomcorbett/OpentokLaravel/Facades/OpenTokApi.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace Tomcorbett\OpentokLaravel\Facades;
-
-use Illuminate\Support\Facades\Facade;
-
-class OpentokApi extends Facade {
-
-    protected static function getFacadeAccessor() { return 'OpentokApi'; }
-
-}
