--- src/config/config.php
@@ -1,13 +1,0 @@
-<?php 
-
-return array( 
-    /**
-     * key
-     */
-    'api_key' => 'API_KEY',
-
-    /**
-     * secret
-     */	
-    'api_secret' => 'API_SECRET',
-);
